Move ShhhDontTellAnyone namespace to Fun namespace; Clean up msbuild
