+++ KitchenLib.TestMod/KitchenLib.TestMod.csproj
@@ -13,13 +13,4 @@
 		<ProjectReference Include="..\KitchenLib\KitchenLib.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-	  <Reference Include="BepInEx">
-	    <HintPath>..\..\..\..\Libraries\BepInEx\core\BepInEx.dll</HintPath>
-	  </Reference>
-	  <Reference Include="MelonLoader">
-	    <HintPath>..\..\..\..\Libraries\MelonLoader\MelonLoader.dll</HintPath>
-	  </Reference>
-	</ItemGroup>
-
 </Project>
