+++ KitchenLib/src/Fun/Stolen/UnlockManager.cs
@@ -13,7 +13,7 @@ using UnityEngine;
  * ALL THIS CODE CAME FROM ICEDMILO'S CARD MANAGER
  * MASSIVE THANKS FOR MAKING THIS SO I DONT HAVE TO <3
  */
-namespace KitchenLib.ShhhDontTellAnyone
+namespace KitchenLib.Fun
 {
 	public class UnlockManager : GameSystemBase
 	{
