+++ KitchenLib/src/Fun/TileHighlighter.cs
@@ -9,7 +9,7 @@ using Unity.Entities;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
-namespace KitchenLib.ShhhDontTellAnyone
+namespace KitchenLib.Fun
 {
 	public class TileHighlighter : UpdatableObjectView<TileHighlighter.ViewData>, ISpecificViewResponse
 	{
